Rename v1 engine to b1
This commit changes the 'v1' engine to 'b1' to represent "bolt v1".
