+++ tsdb/engine/engine.go
@@ -1,5 +1,5 @@
 package engine
 
 import (
-	_ "github.com/influxdb/influxdb/tsdb/engine/v1"
+	_ "github.com/influxdb/influxdb/tsdb/engine/b1"
 )
